Creacion de Entidades Facturas e ItemsFactura y sus respectivos DTO

+++ MarketPlace/src/main/java/com/uade/tpo/MarketPlace/entity/ItemsFactura.java
@@ -21,7 +21,7 @@ public class ItemsFactura {
     private int cantidad;
 
     @Column
-    private int monto;
+    private double monto;
 
     @ManyToOne
     @JoinColumn (name = "producto_id")
